pyproject.toml
[project]
name = "chainlit-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.3.0",
    "google-generativeai>=0.8.5",
    "openai-agent>=0.1.0",
    "python-dotenv>=1.1.0",
]
